add camera flip functionality, remove CameraXLivePreviewActivity.kt and edit graphic overlay layout in WorkOutFragment
