--- app/src/main/res/layout/activity_vision_camerax_live_preview.xml
@@ -1,39 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:keepScreenOn="true">
-
-  <androidx.camera.view.PreviewView
-      android:id="@+id/preview_view"
-      android:layout_width="match_parent"
-      android:layout_height="0dp"
-      app:layout_constraintTop_toTopOf="parent"
-      app:layout_constraintBottom_toBottomOf="parent"/>
-
-  <com.example.poseexercise.views.graphic.GraphicOverlay
-      android:id="@+id/graphic_overlay"
-      android:layout_width="0dp"
-      android:layout_height="0dp"
-      app:layout_constraintLeft_toLeftOf="@id/preview_view"
-      app:layout_constraintRight_toRightOf="@id/preview_view"
-      app:layout_constraintTop_toTopOf="@id/preview_view"
-      app:layout_constraintBottom_toBottomOf="@id/preview_view"/>
-
-  <ToggleButton
-      android:id="@+id/facing_switch"
-      android:layout_width="48dp"
-      android:layout_height="48dp"
-      android:layout_gravity="center_vertical"
-      android:background="@layout/toggle_style"
-      android:checked="false"
-      android:textOff=""
-      android:textOn=""
-      app:layout_constraintBottom_toBottomOf="@+id/preview_view"
-      app:layout_constraintEnd_toEndOf="parent"
-      app:layout_constraintStart_toStartOf="parent" />
-
-
-</androidx.constraintlayout.widget.ConstraintLayout>
