+++ app/src/main/res/layout/fragment_workout.xml
@@ -11,21 +11,23 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent">
 
+
         <androidx.camera.view.PreviewView
             android:id="@+id/preview_view"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.0"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintVertical_bias="1.0"/>
+            app:layout_constraintVertical_bias="0.0" />
 
 
         <com.example.poseexercise.views.graphic.GraphicOverlay
             android:id="@+id/graphic_overlay"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
+            android:layout_width="409dp"
+            android:layout_height="620dp"
             app:layout_constraintBottom_toBottomOf="@id/preview_view"
             app:layout_constraintLeft_toLeftOf="@id/preview_view"
             app:layout_constraintRight_toRightOf="@id/preview_view"
@@ -37,6 +39,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/cancel_button"
+            android:visibility="invisible"
             android:textColor="@color/white"
             app:backgroundTint="@color/primaryColor"
             app:layout_constraintBottom_toBottomOf="parent"
@@ -69,6 +72,7 @@
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
             android:text="@string/complete_button"
+            android:visibility="invisible"
             android:textColor="@color/white"
             app:backgroundTint="@color/blueButton"
             app:layout_constraintBottom_toBottomOf="parent"
@@ -86,7 +90,10 @@
             android:layout_height="wrap_content"
             android:importantForAccessibility="no"
             app:srcCompat="@drawable/camera_flip_icon"
-            app:fabCustomSize="50dp"
+            app:pressedTranslationZ="6dp"
+            app:fabCustomSize="60dp"
+            app:elevation="15dp"
+            app:rippleColor="@color/blueButton"
             app:backgroundTint="@color/white"
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
