--- app/src/main/res/layout/toggle_style.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-  <item
-      android:drawable="@drawable/camera_flip_icon"
-      android:state_checked="true"/>
-  <item
-      android:drawable="@drawable/camera_flip_icon"
-      android:state_checked="false"/>
-</selector>
